Remove/lower noisy warnings
Summary:
The remote setup warning is useful for internal debug.

The VarUpdateWithUpdaterOp warning is invalid. There are cases in the Adam optimizer where the Datatypes do not match, eg
```
 [popart:ir] [warning] In VarUpdateOp::setup(), the VarToUpdate has TensorInfo
FLOAT16 [1024]
but the Updater has TensorInfo
FLOAT   [1024]
```
As a result of removing this warning the `VarUpdateWithoutUpdaterOp` had no difference in function to `VarUpdateOp` so I've removed it entirely.

Test Plan: CI

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Subscribers: jamesn

Maniphest Tasks: T37116

Differential Revision: https://phabricator.sourcevertex.net/D43543