Add reshape collapse pattern
Summary:
This is a workaround for a specific case of T37199.

The new pattern revealed a bug in `hostreduce` transform where `postNrepl` would prune a gradient tensor id that is required for the transform. This is similar to the case where anchor's cannot be touched by patterns, so I fixed it in a similar way.

Test Plan: CI + new test

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: fabiant

Maniphest Tasks: T37199

Differential Revision: https://phabricator.sourcevertex.net/D43613