Host Store / Load Ops with transform
Summary:
This is a WIP revised version of the host load / store diff D35497. This does not enable the host load / stream ops in most cases, instead using a session option.

Now adapted to just change the inputs to have host load ops as producers. The pre/post graph stuff doesn't exist any more as we are trying to adapt it to work with the loop carried dependency stuff in T35921

Test Plan: Given this does not affect the IR in general cases, I have not changed any tests, but will add more as the diff progresses.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: fabiant, jamesn, shirazb, markk, evan, hakons, simonl

Maniphest Tasks: T29603

Differential Revision: https://phabricator.sourcevertex.net/D40246