Remove deprecated SessionOption, accumulationReductionType, and deprecated loss grad op output scaling behaviour when replicatedGraphCount > 1 and reduction is ReductionType::Mean
Summary:
A follow-on from D41185. In that diff, a new SessionOption `accumulationAndReplicationReductionType` was added to supersede (1) the SessionOption `accumulationReductionType` and (2) the scaling behaviour of the loss gradient when replication factor > 1 and a loss has reduction type `ReductionType::Mean`.

Backwards compatibility was retained, but the old behaviour was deprecated. This diff just removes that deprecated behaviour.

I will give apps and poptorch appropriate notice about this before landing.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, markk, stephenm

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: sylvainv, stephenm

Maniphest Tasks: T34809

Differential Revision: https://phabricator.sourcevertex.net/D42303