Add AccumulationType::Mean to allow for gradient accumulation mean reduce that shouldn't overflow
Summary:
This will be particularly important for AutomaticLossScaling as the statistics gathered will likely not account for the "headroom" required for the accumulation.

This changes the default behaviour of Mean gradient accumulation to use AccumulationType::Mean. To access the previous behaviour use sessionOption `meanAccumulationReductionStrategy = MeanReductionStrategy::Post`.

Test Plan: CI; I've also run BERT-L phase 1 with this diff.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: shengf, jamesn, fabiant

Maniphest Tasks: T37245

Differential Revision: https://phabricator.sourcevertex.net/D43332