Remove recompute_0_ip_test
Summary:
This just removes the test in question. This is defunct since D42061 means that the inplace ops are no longer being inplaced.

I tried to add the repro in that diff to test the issue, but couldn't get the recomps to work:

{F1323387}

If anyone can see the issue in that, I'd be grateful.

Test Plan: Buildbots.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn

Maniphest Tasks: T35762

Differential Revision: https://phabricator.sourcevertex.net/D43798