GRUGrad Op: ensure correct shape for bias gradients
Summary: This fixes the shape of bias gradients for GRUOp when linear_before_reset attribute is set to 1. In this case there are actually two biases and the tensor layouts expected by poplib are quite different from the default case. Required for accurate Torch support. Closes T37648

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T37648

Differential Revision: https://phabricator.sourcevertex.net/D44266