SGD2Decompose pattern
Summary:
Depends on D44079.

---

To decompose an SGD2ComboOp:
  # Add a state tensor for accl1. Note we use reservedAccl1Prefix in the id, not Accl as this has a special meaning as the SGD1 combined accl+accum.
  # Add a state tensor for accum, if grad acc.
  # If OptimizerReductionType::GradReduce, perform all reduce on grad.
  # If grad acc, accumulate grad into accum.
  # If accumType float16 and accl1Type FLOAT, upcast accum.
  # Update accl1 using the SGD velocity update equation (*).
  # If grad acc, zero the accumulator.
  # Update the weight using the SGD weight update equation (**).

(*) Updating accl1:
Nothing fancy like SGD1, this is the full SGD equation in one go. We
have reused the SGD1 compound scalars with rf=1 (see D44017).

However, for easier implementation and maintenance, we reuse the
two separate ops for implementing the update in SGD1, just allowing
them to happen immediately one after the other. That is,
SGD1AcclUpdateOp(v, w; smm1, swd1) followed by an
AccumulateOp(v, g; dpsf1).

However, for naming clarity we create an SGD2PartialAcclUpdateOp
that extends an SGD1AcclUpdateOp and is exactly equivalent in all
but name. See the header for more details.

(**) Updating the weight:

This step is again exactly equivalent to SGD1, so we reuse the
SGD1VarUpdateOp. Again, for naming clarity, we create an
SGD2VarUpdateOp that extends this op. See the header for more
details.

---

Other changes:
  - ir.prepare runs SGD2Decompose along with other optimizer decomposers.
  - Make Pattern::transferBaseProperties take `const Op * from`

Test Plan:
Parameterise all existing e2e `optimizer_tests` for SGD2. We add to
`optimizer_tests/get_results.hpp` the optimizer-step-emulating
functions for SGD2 too, then parameterise the tests to use both the
SGD1 and SGD2 functions.

Unit test some basic functionality of SGD2Decompose pattern, but
not `apply`. Proper unit testing of this using the IrQuery testutil will
be in a follow-up task T38253.

Reviewers: fabiant, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: fabiant, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Subscribers: jamesn

Maniphest Tasks: T31043

Differential Revision: https://phabricator.sourcevertex.net/D44279