Update to nll loss for all ignored indices
Summary:
This replicates the pytorch behaviour here:

https://github.com/pytorch/pytorch/blob/2df4acd0253f32a91b9a11a4b1e7c8d392a17a35/aten/src/ATen/native/LossNLL.cpp#L122

Note; the comment:
```
// allow NaN result for total_weight_val == 0 case, see #15870
```
refers to the `total_weight_val == 0` case refers to when the total weight value to be updated is zero and in their implementation this results in a `0/0`. It is unclear from https://github.com/pytorch/pytorch/issues/15870#issuecomment-665484728 whether this is intentional.

We are avoiding the equivalent of `input.numel() == 0` in this diff (`numNonIgnoredSamples == 0`)

In the task, this was suggested as a workaround:
```
auto min_1 = opx.graph().addConstant(numNonIgnoredSamples.elementType(), {}, 1, opx.debugContext("const_1"));
opx.graph().setTileMapping(min_1, 0);
popops::maxInPlace( opx.graph(),
                        numNonIgnoredSamples,
                        min_1,
                        prog,
                        opx.debugContext("numNonIgnoredSamples_min"));

```
See the task for further explanation on why this valid.

Test Plan: Added at test for -1 ignore index.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Subscribers: markk, marieanne

Maniphest Tasks: T36441

Differential Revision: https://phabricator.sourcevertex.net/D44433