Add option to have Adam implicitly remove lossScaling
Summary:
By not removing the lossScaling before updating the optimizer state accl1 can be kept in fp16 without risk of underflowing.

This option also has the effect that the optimizer state will be saved/restored scaled. Users will need to be aware of this.

In a later diff, this will be added to Adaptive and SGD2 Optimizers.

Test Plan: CI; New test; Used in BERT training.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, fabiant

Subscribers: evan, shirazb, jamesn, fabiant

Maniphest Tasks: T38221

Differential Revision: https://phabricator.sourcevertex.net/D45244