Annotate according to scheduling task
Summary:
When growing the poprithms schedule::shift Graph, certain structures are
inserted exclusively to accelerate the shifting process (the refinement step).
The shifting process is not always run. This patch makes these annotations be
omitted if there is to be no shifting. It improves compile time:

From:
       [Scheduler] annotateAccumulateOuterFragmentOps           : 37.623621 [s]     :     4 %

To:
       [Scheduler] annotateAccumulateOuterFragmentOps           : 10.194052 [s]     :     1 %

For a BERT model which takes
       Total                                                    : 769.766694 [s]    :   100 %.

So about a 3% compile time improvement.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Maniphest Tasks: T39277

Differential Revision: https://phabricator.sourcevertex.net/D45565