Fix for map::at error in lstm sdk tests
Summary:
This was due to an issue in the auto virtual graph transform. It was looking for a input tensor that wasn't there. This is a bit of an odd fix, but I am unaware of any other operators that have optional inputs that are not created if not present.
I figured out the tile exceptions were due to me compiling in debug mode, and the test running artificial data, which created nans.

Test Plan: Added a test that checks it.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, george

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, evan, george

Subscribers: george

Maniphest Tasks: T38549

Differential Revision: https://phabricator.sourcevertex.net/D44868