Port popART to use a snap::Graph rather than a poplar one
Summary:
This is the first step towards integrating popIR in PopART.

popIR provides a transitional API called `snap` that will more or less mirror the existing Poplar/Poplibs API but will allow a runtime dispatch between a PopIR and a Poplar backend. Currently the default (and only implemented) backend is the Poplar one. A temporary method called `getPoplarGraph` exists in the snap API to get the internal Poplar graph back out again until the snap API is complete.

This diff replaces all instances of `poplar::Graph` with `snap::Graph` and adds calls to `addPoplarGraph` at all the points where we call through to Poplar and Poplibs. Going forward we will need to move all of these places into the snap API and once there is no more calls to `getPoplarGraph` in PopART then it is ready to enable the popIR backend.

To keep existing custom ops from breaking the `Opx` interface has not changed and this diff adds a new interface called `PopOpx`. The original `Opx` is now implemented in terms of `PopOpx` and all internal ops have been ported to the new API.

The `Opx` API will be deprecated in 2.2 (not this release because there is not currently an alternative) and removed in 2.3.

Test Plan: all existing tests still pass including (importantly) the custom op example in the docs folder.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, jakeh, stephenm, georgep, george, tomm, fabiant, markk, evan, shirazb, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, markk

Subscribers: shirazb, jamesn

Maniphest Tasks: T35454

Differential Revision: https://phabricator.sourcevertex.net/D45304