Do not save accum__ tensor in serialised executable.
Summary:
T36910 was running into an issue where the saved executable did not find the Accum__ tensors. These should not be stored and thus should not be added to `AdditionalModelProtoTensors`. This just removes this adding.

I also corrected a couple of things.

Test Plan: Added test in executable_serializaition_test.cpp which fails on master, passes with this test.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, jamesn, fabiant, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, shirazb

Maniphest Tasks: T36910

Differential Revision: https://phabricator.sourcevertex.net/D45582