Make code from SubgraphOutline transform accessible for use elsewhere
Summary:
SubgraphOutline is a transform that iteratively creates re-usable subgraphs and replaces ops in a Graph with calls to those subgraphs.
There is a lot of code in this transform that we may want to use elsewhere, such as in the explicit pipelining transform.

This diff does not make and functional changes, but makes this code useable outside of subgraphoutline.cpp:
- Make `createSubgraph` a static function in the `SubgraphOutline` class
- Make `replaceWithCallOp` a static function in the `SubgraphOutline` class
- Expose the `Instance` class, and give it the more descriptive name `SubgraphableOpCluster`

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Subscribers: matthewha

Maniphest Tasks: T35924

Differential Revision: https://phabricator.sourcevertex.net/D46036