Do not retain sourceIpus and sourceTensors state when cloning an IpuCopyOp
Summary:
A cloned IpuCopyOp is not connected, but currently its state, namely `sourceTensors` and `sourceIpus` members, are copied to the clone.

This causes issues when e.g. `IpuCopyOp::copiesOptimizerTensors()` is called on the clone, which iterates through `sourceIpus`. If a tensor of the clonee's `sourceIpus` map doesn't exist in the graph of the cloned op, then you see a `IndexError: map::at` exception.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Maniphest Tasks: T35924

Differential Revision: https://phabricator.sourcevertex.net/D46122