Move PipelineInfo class into Ir level for use in explicit pipeline IR transform
Summary: The `PipelineInfo` class declaration has been moved from irlowering.hpp to pipeline.hpp (the Ir transform). This is a more sensible home for it, as it can be used in the explicit pipelining transform, and will eventually no longer be needed in IrLowering. No functional changes.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Maniphest Tasks: T35924

Differential Revision: https://phabricator.sourcevertex.net/D46222