Handle no ONNX model in Ir::serialise
Summary:
Fixes T39543.

We use `graph->id` as the graph's key in the json.

Like before, the maingraph name becomes the name of the maingraph in the onnx
model, if one exists; HOWEVER previously we only did this when the maingraph was
the first graph encountered, but now this always happens whether the maingraph
was first or not.

**Thus there is a subtle api change when `useScheduler=false` is passed.**

Test Plan: Add deonnxing regression unit tests for all cases.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george

Maniphest Tasks: T39543

Differential Revision: https://phabricator.sourcevertex.net/D50037