Fix fp16 optimizer state with non-const lossScaling
Summary: OptimizerTensors are always fp32. In the case where any of the Optimizer**State**Tensors in Adam are not fp32 and the weight is fp16, the gradient to be unscaled will be fp16. This causes a type mismatch with non-const lossScaling. This patch adds a castOp to resolve.

Test Plan: CI; New unittest.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T38425

Differential Revision: https://phabricator.sourcevertex.net/D46488