Correct dtype in Pooling ops
Summary:
See D46510 . Before this diff the partials type was always in FP16; the diff to add the code I have removed was added in D8296 and seems to have had no effect. Amazingly, it seems nobody had tested an averagepool with float 16 before now.

The option to have partials in fp32 remains a poplar engine option for the reasons discussed in D46510. We can add an option to enable later if desired.

Resolves T38924

Test Plan: Added test and tested on example in T38924

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk

Maniphest Tasks: T38924

Differential Revision: https://phabricator.sourcevertex.net/D46639