Use embedding planner in ScatterReduceGrad
Summary:
This change is necessary for optimising the performance of a simple graph neural net example from PyTorch geometric.  The attached screenshot shows a comparison profile report (left is the latest sdk, right is a build incorporating this diff).  Note that the number of cycles goes from 7186633 to 533885 with this diff.  More importantly, the embedding planner allows running GNNs with larger graphs without running out of memory.

 {F1444398}

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Maniphest Tasks: T40427

Differential Revision: https://phabricator.sourcevertex.net/D47300