Clip the loss scale at max(fp16) if the loss scale is in fp32 and any of the weights are fp16.
Summary:
Add an option to `LossScaleUpdateOp` to clip the final loss at the largest power of 2 that fits in `fp16` -- 32768.

Set that option to true in the automatic loss scaling transform if either of these is true:
- The loss scale tensor is in `fp16`.
- Any of the tracked tensors are in `fp16`.

An edge case that might cause the above to break is when `lossScale < 1` (for example 0.5) and the type of loss scale is `fp16`. In that case when the final loss scale - that gets applied to the gradients - is `finalLossScale = 0.5 * factor = 2^15`, the `factor` - also `fp16` - will be expected to be `2^16`, but it will be `max(fp16)`, which is a slightly lower number.

Fixes T38779.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, romand, oskarb, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oskarb, matthewha

Subscribers: arjunc, jamesn, romand, matthewha

Maniphest Tasks: T38779

Differential Revision: https://phabricator.sourcevertex.net/D46352