Create PreAutomaticLossScale transform.
Summary:
This transform reads a user-specified list of tensor IDs (provided as `SessionOptions().automaticLossScalingSettings.toTrackTensors`) and inserts an `AutomaticLossScalingProxyOp` between each of these tensors and its destination. The motivation is that `Autodiff` will add the corresponding `AutomaticLossScalingProxyGradOp` in the backward graph, so that we can find the gradients of the user-specified list of tensor IDs when doing automatic loss scaling.

Example graph before applying the transform:
```
A -- MulOp -- C
B -'
```

Example graph after applying the transform with toTrackTensors = ["A", "C"]:
```
A -- AlsProxyOp -- A* -- MulOp -- C -- AlsProxyOp -- C*
B ---------------------'
```

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, romand, markk, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Subscribers: matthewha

Maniphest Tasks: T40309

Differential Revision: https://phabricator.sourcevertex.net/D46718