Refactor SGD0 to have a combo Op and support gradient accumulation
Summary:
This change brings SGD0 inline with other optimizers and uses the OptimizerDecompose methods for constructing replication and gradient accumulation. This allows for changes to optimizer decompose, such as D47403, to effect all optimizers.

The SGD variants will now be selected as the following:
```
SGD0 is No Momentum
SGD1 is Momentum and SGDAccumulatorAndMomentum::Combined
SGD2 is Momentum and SGDAccumulatorAndMomentum::Separate
```
Changes to the graphs will be:
* `SGD0VarUpdate` will be used instead of `SGD1VarUpdate`.
Changes to the numerics:
* The gradient accumulation will now be scaled by `lossScaling`
* The gradient accumulation will no longer be scaled by `replicationFactor`

Test Plan: CI; New sgd0 variants in tests.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, shirazb

Subscribers: matthewha, shirazb

Maniphest Tasks: T41866

Differential Revision: https://phabricator.sourcevertex.net/D47883