Script to add info on each op constructor to json
Summary:
A lot of trial and error in this one. Seems to output the correct thing on my machine; getting it working as part of builds will be difficult.

It was really slow initially so I had to limit recursion and add multi threading. I may live to regret that.

This is the output it generates:

{F1574815}

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, shirazb

Subscribers: shirazb, markk, yanislavd, george

Maniphest Tasks: T41773, T41718

Differential Revision: https://phabricator.sourcevertex.net/D47823