Fix potential bug in `LSTMGradOp::gradInputInfo`.
Summary: gradInputInfo returned a const ref to a static variable local to the function, but the function would modify this variable depending on the forward op. This could lead to the result being incorrect for some ops. I have change the static variable to be a member of LSTMGradOp, allowing for different ops to have different results for calls to `gradInputInfo`.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T39738

Differential Revision: https://phabricator.sourcevertex.net/D48931