Add fix for maxWeightNorm == 0 and allow weight specific versions.
Summary:
The previous behaviour was incorrect when `maxWeightNorm` was set to 0. The fallback condition should be `φ(norm(w)) == 0 or norm(u) == 0`, where `φ` is `min(maxWeightNorm, ...)`. But the previous behaviour was failing the fallback condition when `maxWeightNorm` was set to `0`.

Allows for weight specific `maxWeightNorm`. For consistency the user should now provide `defaultMaxWeightNorm` in the dictionary with the optimizer. `maxWeightNorm` can still be provided but it will raise a deprecation warning.

Test Plan: CI;

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george

Maniphest Tasks: T43224

Differential Revision: https://phabricator.sourcevertex.net/D48769