Add SessionOption to undo late scheduling of VarUpdateOps that are not for weights but consume gradients
Summary:
When recomputation is on, often it is beneficial to schedule the var updates later, but the scheduler will not find this due to not knowing about implicit recomputation. Hence, the `delayVarUpdates` SessionOption was added. However, for variables that are not weights that consume gradients, this is the opposite behaviour from what we want. Consider the `AccumulateOp` for gradient accumulation; we want to consume and then recycle the memory of the gradient as early as possible. Similarly for the `HistogramOps` added during automatic loss scaling.

Thus this diff makes the following changes:

  # Add `scheduleNonWeightUpdateGradientConsumersEarly ` member to `SessionOptions`, and add to hash calculation.
  # `OptimizerDecompose::gradAccum` overrides `AccumulateOp's` schedule priority to max if `shouldDelayVarUpdates` (meaning no explicit recompute and main loops) and `scheduleNonWeightUpdateGradientConsumersEarly `.
  # ALS transform overrides `HistogramOp`s schedule priority to max if `shouldDelayVarUpdates` (meaning no explicit recompute and main loops) and `scheduleNonWeightUpdateGradientConsumersEarly `.

The `AccumulateOp` gets its base properties from the combo op which, as a `VarUpdateOp`, will have -inf schedule priority if `shouldDelayVarUpdates`. However, gradient accumulation `AccumulateOps` are not for updating weights; they consume gradients. We want to consume (and thus be able to recycle the memory of) the gradients as soon as possible, and therefore we re-override their schedule priority to +inf. A similar line of reasoning is made for the `HistogramOp`s.

We also add python bindings for the two SessionOptions.

Aside from the above change, in the optimizerdecompose unit test, remove the addTensor method and inline its uses, as was resulting in pointer to memory on its destroyed stack frame. Secondly, in cmake, do not link test graphs test util for this test, as it is not used.

Test Plan:
unit test `OptimizerDecompose::gradAccum`.
Set the option on an existing ALS integration test and inspect the Ir for the schedule priority of the Histogram ops.

Reviewers: oskarb, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, yanislavd

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, yanislavd

Subscribers: matthewha

Maniphest Tasks: T43302

Differential Revision: https://phabricator.sourcevertex.net/D48811