added shape dimension check for when no InputShapeTensor is provided
Summary:
A customer ran into an issue importing an ONNX file with non-specified shape dimensions (which is not supported in PopART as you must use the `InputShapeInfo` API). PopART expects these unknown dimension cases to be set as a `dim_param` [(string) value](https://github.com/onnx/onnx/blob/master/onnx/onnx.proto#L655) and so has logic to catch these and throw a useful error to the user but in this case, the unknown dimensions were set as negative values which causes unrelated errors later in the PopART IR processing. I believe the -ve values may denote "expanding" dimensions (although I was unable to verify that in the ONNX documentation). Even so, I don't see how that could be supported by PopART so this case should be handled the same way as `dim_param` values.
This diff adds logic to handle such -ve dim cases in the same way and includes a new test case for them.

Test Plan: CI

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, michaell

Subscribers: michaell, jamesn

Maniphest Tasks: T42275

Differential Revision: https://phabricator.sourcevertex.net/D48125