Add Support for Transpose on both sides of TiedGather
Summary:
Generalises TiedGatherPattern to work with a transpose between the `root_weight` and `matmul`.

Added a new tguil method `maybeTraverseProducer`

---

PyTorch BERT was developed based on a bug in DetachOp that caused the tile
layout of the weight to be determined by the MatMul, not the TiedGather.
This costs extra cycles due to the exchange, but actually has lower
memory footprint.

In order to natively support PyTorch BERT in the short-term, we thus have to
disable the TiedGather input creator, so we get the same behaviour that
resulted from the DetachOp bug.

Test Plan: CI; Variant in `testtiedgather.py`; Ran PyTorch BERT

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oskarb, fabiant, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, oskarb, fabiant, matthewha

Maniphest Tasks: T28669

Differential Revision: https://phabricator.sourcevertex.net/D48926