Only create new pass through tensor if not already present.
Summary: This resolved a segfault issue that could be caused by transforms which cloned ops. If the cloner cloned an lstm op, then connected inputs and outputs to matched the original op, and then called setup on the clone, `LSTMOp::setup` would disconnect some of these outputs and replace them. The outputs the cloner now thought were connected to the lstm clone, are actually tensors with no producer.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T44434

Differential Revision: https://phabricator.sourcevertex.net/D49988