Fix segfault issue in ReverseOpx.
Summary: There was a dynamic_cast casting to `ReverseOp` when it should have been using `ReverseBaseOp`. This caused the segfault when op_p was a `ReverseInplaceOp`. I've changed the cast to use `ReverseBaseOp`, but also changed uses of `dynamic_cast` to `PopOpx::getOp` as this also checks the cast was successful.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, jamesn

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, markk, jamesn

Maniphest Tasks: T44543

Differential Revision: https://phabricator.sourcevertex.net/D50162