Add base Autofix linter for reliably applying lint patches
Summary:
This diff contains the base classes for the python linter orchestration. Other concrete subclass linter implementations will depend on this, one by one.

The idea is to provide a python implementation or wrapper for linters which apply autofixes. That is, linters which 'suggest' changes, as opposed to just flagging errors.

The need for this comes from the fact that `arc lint` does not accumulate patches from multiple linters into a single patch to be applied to a file. For example, suppose we run linter A and linter B on `some_file`, and both
linters produce a patch to be applied. In arc, when we say `Y` to apply the patch, this will only be from that last linter to run, so we lose the patch from either linter A or litner B (depending on which was first) and we're
forced to run arc lint again, or indefinitely if linters end up correcting each others patches.
This python linter orchestrator will accumulate linter patches into a single patch which then gets passed back to arc lint, avoiding this issue as arc only ever sees one patch which is the unification of changes from multiple litners.

Test Plan:
Added tests specific to the linters.
Tested separately as a script and also with `arc lint`

Reviewers: yanislavd, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Reviewed By: yanislavd, #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved

Maniphest Tasks: T43947

Differential Revision: https://phabricator.sourcevertex.net/D50356