Let ParsedTensorId parser search through the Ir for scopes
Summary:
In order to sanitise the scopes in a `TensorId`, this diff enables the `ParsedTensorId` to search through the `Ir` for valid scopes.

- In order to acquire the `Ir` the `addScope` and `removeScope` in `util.{cpp,hpp}` is accepting a `Graph` object (which holds both the `Scope` and the `Ir`) instead of a `Scope` object.
- The constructor of `ParsedTensorId` requires an `Ir` object
- The parsing of `Scopes` no longer splits by `sNameDelimeter`, but instead searches for valid `Scopes` found in the `Ir` (as such the user is still free to name tensors like `MatMul:0/2` without `MatMul:0` being a scope)

Limitations:
It's assumed that the scopes are in the beginning of the `TensorId`. An ill formatted `TensorId` like `foo_myScope_bar` will be returned as `myScope/bar` without throwing a warning.

Test Plan: Updated unittests to reflect the changes

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george, fabiant

Subscribers: fabiant, george

Maniphest Tasks: T45640

Differential Revision: https://phabricator.sourcevertex.net/D51485