Combine the user-defined loss scale value with other optional scalings on the host before streaming to device. This avoids the need for an extra 'Scale' operation in the graph
Summary:
When doing the autodiff transform in popart, the loss scale tensor is passed as the 'grad in' tensor to the gradient op of the op in the forward pass that produces the loss tensor (let's call this 'loss grad op').

The scalar loss scale value is set by the user. Under certain conditions*, this loss scale is scaled by an additional factor before being passed as input to the loss grad op. Currently this scaling occurs on the device:

```loss_scale -> ScaleOp -> final_loss_scale -> LossGradOp```

This diff moves this scaling onto the host, so that when the user sets or updates the optimizer, the value that is streamed to the loss scale tensor on the device is the final loss scale value. This removes the need for the extra on-device scaling.

A precursor diff to D51967, as part of T44406

* The conditions:
```replicatedGraphCount > 1 && opts.accumulationAndReplicationReductionType == ReductionType::Mean && opts.meanAccumulationAndReplicationReductionStrategy == MeanReductionStrategy::PostAndLoss;```

Test Plan: No functionality change, just refactoring, so existing CI test coverage is sufficient.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, romand, markk

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, romand

Maniphest Tasks: T44406

Differential Revision: https://phabricator.sourcevertex.net/D52016