Adding zsh compatible backup to bash only var
Summary:
Adding a backup Zsh compatible equivalent to the $BASH_SOURCE[0] variable which is only defined by default in bash environments. With this, and changes made in D49438, popsdk will be able to setup and activate poplar and popart environments in zsh as well as bash (and potentially other shells in the future).
More desciption and discussion in T43422.

Will add a reviewer as soon as I figure out who I can ask to review this.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george, jorgec, shirazb

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, george, jorgec, shirazb

Subscribers: jorgec, shirazb, george

Maniphest Tasks: T43422

Differential Revision: https://phabricator.sourcevertex.net/D49692