Fix DynamicSlice for axis_size % slice_size != 0
Summary:
In the previous implementation `DynamicSlice` would fail if the size of the axis was not divisible by the size of the slice.
Furthermore, the resulting tile layout of the slice was dependent on the index.

This diff fixes the divisibility constraints and ensures that the slices from `createSliceableTensorFromSlice` have identical layout.
The slices will be spread across fewer tiles, but huge exchange copies is avoided as the output layout does not depend on the index
In addition, this diff updates documentation for `DynamicBaseOp`

Test Plan: Test which would have failed in the previous implementation added.

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T46698

Differential Revision: https://phabricator.sourcevertex.net/D52268